// src/api/generated/models/getCompaniesCompanyIdQuotesDirection.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdQuotesDirection = typeof GetCompaniesCompanyIdQuotesDirection[keyof typeof GetCompaniesCompanyIdQuotesDirection];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdQuotesDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;