// src/api/generated/models/tradingPartnerStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type TradingPartnerStatus = typeof TradingPartnerStatus[keyof typeof TradingPartnerStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TradingPartnerStatus = {
  INVITED: 'INVITED',
  ACTIVE: 'ACTIVE',
  REJECTED: 'REJECTED',
} as const;